[runSofa] Add a button and a mechanism to activate the scenegraph updates (#3026)
The scene graph view is not updated to reflect the content of the scene changes. 
This is because a mecanism was implemented to not update the view at each simulation step to maximise performance. 

This may be useful when max performance is needed but this lead to an inconsistant display. 

In this PR I try to fix the situation by allowing three states:
- lock = true => changes in the scene graph are not reflected into the model but the model is flagged as dirty 
- lock = false => changes in the scene graph are imediately reflected into the model
- dirty => true => the model is not anymore synchronized with the sofa scene graph
- dirty => false => the model is showing the content of the scene graph

when going from lock=true to lock=false the model is updated if it was dirty. 

To let user to control if they want the graph to be updated we add an icon to lock/unlock the model state. If the model is dirtyfied
a the icon is reflecting the change (with an highly professional icon free of copyright ;)) 

Fixes #2920 


The PR also contains few fixes for the modeller (that is still broken anyway).